# optimize stopping strings processing