@mcmonkey4eva 2023-04-28
There's actually also a related but separate issue raised with this code (irrelevant to the PR, same issue before/after):

If beam search is used, this is designed such that *any match* in any beam stops the entire run, rather than only forbidding beams that matched.

In practice this means that higher beam count == higher chance of unexpected stops, as if any beam generates a stop string, the whole gen is stopped.

This came up on Discord as `DeSinc` had beam search = 3 and the bot would cut off replies mid-sentence for no reason, which we traced back to one of the beams generating a stop string, and the others not. With beam search = 1 it just doesn't cut off at all.

Not sure how to address that, as the relevant implementation code is deeper into internals than I'd like to venture.

---

That's not a blocker on the PR though, entirely separate topic to look into, just came up while looking into this code.