@oobabooga 2023-05-02
Very good idea. One of the best kinds of PR: removes lines of code, makes things faster.